// src/types/contentful.d.ts
// THIS FILE IS AUTOMATICALLY GENERATED. DO NOT MODIFY IT.

import { Asset, Entry } from "contentful";
import { Document } from "@contentful/rich-text-types";

export interface IHomeFields {
  /** Заголовок  */
  title?: string | undefined;

  /** Баннер  */
  banner?: Asset | undefined;

  /** Гарнизон Это ... */
  infoText?: Document | undefined;
}

export interface IHome extends Entry<IHomeFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "home";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}

export interface ILinksFields {
  /** Instagram */
  instagram?: string | undefined;

  /** Telegram */
  telegram?: string | undefined;

  /** Номер телефона */
  phone?: string | undefined;

  /** WhatsApp */
  whatsApp?: string | undefined;
}

export interface ILinks extends Entry<ILinksFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "links";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}

export interface IPlansFields {
  /** Название */
  title?: string | undefined;

  /** Цена */
  price?: string | undefined;

  /** Популярный */
  isPopular?: boolean | undefined;

  /** Кол-во вызовов */
  callsQuantity?: string | undefined;
}

export interface IPlans extends Entry<IPlansFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "Plans";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}

export interface IServicesFields {
  /** Заголовок */
  title: string;

  /** slug */
  slug?: string | undefined;

  /** Описание */
  description?: Document | undefined;

  /** Картинка */
  photo: Asset;
}

export interface IServices extends Entry<IServicesFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "services";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}

export type CONTENT_TYPE = "home" | "links" | "Plans" | "services";

export type LOCALE_CODE = "en-US";

export type CONTENTFUL_DEFAULT_LOCALE_CODE = "en-US";